challange1a/requirements.txt
PyPDF2==3.0.1
pdfminer.six==20231228
